Preparation for transaction v0 work.
No logical changes

Moved some stuff around so the structs are better organized
